Prototype navigation ui
The ui requires a seperate branch with unique templates to allow
simpler prototyping.  This branch should only contain changes
relevant to the user interface. Seperating the ui development also
allows isolation from breaking changes down the line. Once a broken
interface is stable, it may be merged with the main branch.

NOTE: This is a guideline, not a rule... but it helps to follow it.